% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{add_account}
\alias{add_account}
\title{Agregar Cuenta de Twitter}
\usage{
add_account(
  username,
  password,
  email,
  email_password,
  cookies = NULL,
  db_file = "accounts.db"
)
}
\arguments{
\item{username}{Nombre de usuario}

\item{password}{Contraseña}

\item{email}{Email de la cuenta}

\item{email_password}{Contraseña del email}

\item{cookies}{Cookies de sesión (opcional pero recomendado)}

\item{db_file}{Archivo de base de datos (default: "accounts.db")}
}
\description{
Agrega una cuenta de Twitter para usar en scraping
}
\examples{
\dontrun{
add_account(
  username = "mi_usuario",
  password = "mi_password",
  email = "mi@email.com",
  email_password = "email_pass",
  cookies = "auth_token=...; ct0=..."
)
}
}
