% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.data.frame.twscraper_user}
\alias{as.data.frame.twscraper_user}
\title{Convertir Usuario a Data Frame}
\usage{
\method{as.data.frame}{twscraper_user}(x, ...)
}
\arguments{
\item{x}{Usuario de get_user()}

\item{...}{Argumentos adicionales}
}
\value{
data.frame
}
\description{
Convierte un usuario individual a data.frame de una fila
}
