% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accounts.R
\name{delete_account}
\alias{delete_account}
\title{Eliminar Cuenta}
\usage{
delete_account(username, db_file = "accounts.db")
}
\arguments{
\item{username}{Nombre de usuario a eliminar}

\item{db_file}{Archivo de base de datos (default: "accounts.db")}
}
\description{
Elimina una cuenta de la base de datos
}
\examples{
\dontrun{
delete_account("mi_usuario")
}
}
