% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\docType{data}
\name{.twscraper_env}
\alias{.twscraper_env}
\title{Entorno global para almacenar configuración}
\format{
An object of class \code{environment} of length 3.
}
\usage{
.twscraper_env
}
\description{
Entorno global para almacenar configuración
}
\keyword{internal}
