% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_by_lang}
\alias{filter_by_lang}
\title{Filtrar Tweets por Idioma}
\usage{
filter_by_lang(tweets, lang)
}
\arguments{
\item{tweets}{Lista de tweets}

\item{lang}{Código de idioma (ej: "es", "en", "pt")}
}
\value{
Lista filtrada de tweets
}
\description{
Filtra tweets por idioma
}
\examples{
\dontrun{
tweets <- search_tweets("Milei", n = 100)
tweets_es <- filter_by_lang(tweets, "es")
}
}
