% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{get_user}
\alias{get_user}
\title{Obtener Información de Usuario}
\usage{
get_user(username, progress = TRUE)
}
\arguments{
\item{username}{Nombre de usuario (sin @)}

\item{progress}{Mostrar progreso (default: TRUE)}
}
\value{
Lista con información del usuario
}
\description{
Obtiene información detallada de un usuario
}
\examples{
\dontrun{
user <- get_user("hadleywickham")
}
}
