% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.twscraper_tweet}
\alias{print.twscraper_tweet}
\title{Método Print para Tweet Individual}
\usage{
\method{print}{twscraper_tweet}(x, ...)
}
\arguments{
\item{x}{Información de un tweet}

\item{...}{Argumentos adicionales}
}
\description{
Imprime información de un tweet individual
}
