% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.twscraper_tweets}
\alias{print.twscraper_tweets}
\title{Método Print para Tweets}
\usage{
\method{print}{twscraper_tweets}(x, ...)
}
\arguments{
\item{x}{Lista de tweets}

\item{...}{Argumentos adicionales}
}
\description{
Imprime resumen de tweets
}
