% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.twscraper_user}
\alias{print.twscraper_user}
\title{Método Print para Usuario}
\usage{
\method{print}{twscraper_user}(x, ...)
}
\arguments{
\item{x}{Información de usuario}

\item{...}{Argumentos adicionales}
}
\description{
Imprime información de usuario
}
