% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.twscraper_users}
\alias{print.twscraper_users}
\title{Método Print para Lista de Usuarios}
\usage{
\method{print}{twscraper_users}(x, ...)
}
\arguments{
\item{x}{Lista de usuarios}

\item{...}{Argumentos adicionales}
}
\description{
Imprime resumen de usuarios
}
