% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_csv}
\alias{save_csv}
\title{Guardar Tweets como CSV}
\usage{
save_csv(tweets, file)
}
\arguments{
\item{tweets}{Lista de tweets}

\item{file}{Nombre del archivo}
}
\description{
Guarda tweets en formato CSV
}
\examples{
\dontrun{
tweets <- search_tweets("rstats", n = 100)
save_csv(tweets, "tweets_rstats.csv")
}
}
