% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_json}
\alias{save_json}
\title{Guardar Tweets como JSON}
\usage{
save_json(tweets, file)
}
\arguments{
\item{tweets}{Lista de tweets}

\item{file}{Nombre del archivo}
}
\description{
Guarda tweets en formato JSON
}
\examples{
\dontrun{
tweets <- search_tweets("rstats", n = 100)
save_json(tweets, "tweets_rstats.json")
}
}
