% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sort_tweets}
\alias{sort_tweets}
\title{Ordenar Tweets}
\usage{
sort_tweets(tweets, by = "date", decreasing = TRUE)
}
\arguments{
\item{tweets}{Lista de tweets}

\item{by}{Campo por el cual ordenar ("date", "like_count", "retweet_count", "views_count")}

\item{decreasing}{Orden descendente (default: TRUE)}
}
\value{
Lista ordenada de tweets
}
\description{
Ordena tweets por algún criterio
}
\examples{
\dontrun{
tweets <- search_tweets("Milei", n = 100)
top_tweets <- sort_tweets(tweets, by = "like_count")
}
}
