% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_dataframe}
\alias{to_dataframe}
\title{Convertir Tweets o Usuarios a Data Frame}
\usage{
to_dataframe(x)
}
\arguments{
\item{x}{Lista de tweets o usuarios}
}
\value{
data.frame o tibble
}
\description{
Convierte una lista de tweets o usuarios a data.frame usando purrr::map_dfr
}
\examples{
\dontrun{
tweets <- search_tweets("rstats", n = 100)
df <- to_dataframe(tweets)

users <- get_followers("rstudio", n = 100)
df <- to_dataframe(users)
}
}
