% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping_advanced.R
\name{users_to_dataframe}
\alias{users_to_dataframe}
\title{Convertir Usuarios a Dataframe (alias de to_dataframe)}
\usage{
users_to_dataframe(users)
}
\arguments{
\item{users}{Lista de usuarios}
}
\value{
data.frame con información de usuarios
}
\description{
Alias de to_dataframe() para retrocompatibilidad
}
\examples{
\dontrun{
followers <- get_followers("elonmusk", n = 100)
df <- users_to_dataframe(followers)
# Es equivalente a:
df <- to_dataframe(followers)
}
}
